Himmelblau requires a mechanism to send pam_prompt
In Himmelblau I need a mechanism for sending an
additional pam prompt (for performing 2FA via
a browser window on another device). I've
implemented that here by providing a pam continue
feature, which sends a message with some optional
data back to pam for an additional prompt. The
purpose of the additional data is simply for
continuing the authentication later (in my case,
it's a json string).

Signed-off-by: David Mulder <dmulder@samba.org>